+++ packages/form-table/src/index.vue
@@ -195,6 +195,10 @@ defineExpose({
 
 <style lang="less" scoped>
 .form-table-container {
+  :deep(.form-table-field-layout) {
+    flex-wrap: wrap;
+  }
+
   :deep(.el-table__cell .el-form-item) {
     margin-bottom: 0;
   }